requirements.txt
Flask
sass
Markdown
python-geoip-python3
Jinja2
pyyaml
uwsgi
coverage
